#82 [Fix] Timer 수정
